enhancement (core): rename date file
