+++ projects/knora/core/src/lib/declarations/index.ts
@@ -13,7 +13,7 @@ export * from './api/knora-constants';
  * Interfaces for shared objects
  */
 export * from './api/shared/strings';
-export * from './api/shared/shared-date';
+export * from './api/shared/date';
 
 /**
  * Interfaces for authentication
